requirements.txt
asgiref==3.2.3
astroid==2.3.2
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.2
django-cors-headers==3.1.1
django-debug-toolbar==2.1
django-extra-fields==2.0.0
django-filter==2.2.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-taggit==1.1.0
django-taggit-serializer==0.1.7
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangorestframework-xml==1.4.0
drf-yasg==1.17.0
gunicorn==20.0.4
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.17.3
openapi-codec==1.3.2
packaging==19.2
pandas==0.25.3
pep8==1.7.1
Pillow==6.2.1
psycopg2==2.8.4
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.3
pyparsing==2.4.5
python-dateutil==2.8.1
python-decouple==3.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
simplejson==3.16.0
six==1.13.0
sqlparse==0.3.0
static3==0.7.0
text-unidecode==1.3
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
wrapt==1.11.2
